+++ SmartCar Core/src/main/java/be/uantwerpen/sc/tools/Dijkstra.java
@@ -18,7 +18,7 @@ public class Dijkstra
         List<Vertex> vertexList = new ArrayList<>();
         while (!vertexQueue.isEmpty()) {
             Vertex u = vertexQueue.poll();
-            Vertex v = new Vertex(1);
+            Vertex v = new Vertex(1L);
             // Visit each edge exiting u
             for (Edge e : u.getAdjacencies())
             {
@@ -55,14 +55,14 @@ public class Dijkstra
         vertex= vertexes.get(i);*/
 
 
-        int i = target.getId();
+        int i = (int)(long)target.getId();
         do {
             for (Vertex v : vertexes){
                 if(v.getId() == i)
                     path.add(vertexes.get(i-1));
             }
             try {
-                i = vertexes.get(i-1).getPrevious().getId();
+                i = (int)(long)vertexes.get(i-1).getPrevious().getId();
             }catch (Exception e){
                i = 0;
             }
